feat(dialog): move and rename NewConversationContactsPage
