+++ src/components/ContactSelectionBubble.vue
@@ -46,9 +46,9 @@ import Close from 'vue-material-design-icons/Close.vue'
 
 import NcButton from '@nextcloud/vue/dist/Components/NcButton.js'
 
-import AvatarWrapper from '../../../AvatarWrapper/AvatarWrapper.vue'
+import AvatarWrapper from './AvatarWrapper/AvatarWrapper.vue'
 
-import { AVATAR } from '../../../../constants.js'
+import { AVATAR } from '../constants.js'
 
 export default {
 	name: 'ContactSelectionBubble',
